feat(j-s): Semi optional ruling (#7525)
* Make ruling optional in restriction cases

* Remove validation from ruling in restriction cases

* Make ruling optional in investigation cases

* Move courtRecordForm into courtRecord

* Fix e2e tests

* Revert infra change
