+++ apps/judicial-system/web-e2e/src/integration/Court/InvestigationCases/courtRecord.spec.ts
@@ -143,48 +143,6 @@ describe(`${IC_COURT_RECORD_ROUTE}/:id`, () => {
     cy.getByTestid('sessionBookings').should('not.match', ':empty')
   })
 
-  it('should require a valid court location', () => {
-    const caseData = makeInvestigationCase()
-
-    const caseDataAddition: Case = {
-      ...caseData,
-      prosecutor: makeProsecutor(),
-      courtDate: '2021-12-16T10:50:04.033Z',
-    }
-
-    cy.stubAPIResponses()
-    cy.visit(`${IC_COURT_RECORD_ROUTE}/test_id_stadfest`)
-
-    intercept(caseDataAddition)
-
-    cy.wait('@gqlCaseQuery')
-    cy.getByTestid('courtLocation').clear().blur()
-    cy.getByTestid('inputErrorMessage').contains('Reitur má ekki vera tómur')
-    cy.getByTestid('courtLocation').type(faker.lorem.word())
-    cy.getByTestid('inputErrorMessage').should('not.exist')
-  })
-
-  it('should require valid session bookings', () => {
-    const caseData = makeInvestigationCase()
-
-    const caseDataAddition: Case = {
-      ...caseData,
-      prosecutor: makeProsecutor(),
-      courtDate: '2021-12-16T10:50:04.033Z',
-    }
-
-    cy.stubAPIResponses()
-    cy.visit(`${IC_COURT_RECORD_ROUTE}/test_id_stadfest`)
-
-    intercept(caseDataAddition)
-
-    cy.wait('@gqlCaseQuery')
-    cy.getByTestid('sessionBookings').clear().blur()
-    cy.getByTestid('inputErrorMessage').contains('Reitur má ekki vera tómur')
-    cy.getByTestid('sessionBookings').type(faker.lorem.words(5))
-    cy.getByTestid('inputErrorMessage').should('not.exist')
-  })
-
   it('should not allow users to continue if conclusion is not set', () => {
     const caseData = makeInvestigationCase()
 
@@ -193,6 +151,7 @@ describe(`${IC_COURT_RECORD_ROUTE}/:id`, () => {
       prosecutor: makeProsecutor(),
       courtDate: '2021-12-16T10:50:04.033Z',
       decision: CaseDecision.ACCEPTING,
+      conclusion: undefined,
     }
 
     cy.stubAPIResponses()
@@ -200,7 +159,10 @@ describe(`${IC_COURT_RECORD_ROUTE}/:id`, () => {
 
     intercept(caseDataAddition)
 
-    cy.getByTestid('continueButton').should('not.exist')
+    cy.getByTestid('formFooter')
+      .children()
+      .getByTestid('infobox')
+      .should('exist')
   })
 
   it('should not allow users to continue if decision is not set', () => {
@@ -211,6 +173,7 @@ describe(`${IC_COURT_RECORD_ROUTE}/:id`, () => {
       prosecutor: makeProsecutor(),
       courtDate: '2021-12-16T10:50:04.033Z',
       conclusion: faker.lorem.words(5),
+      decision: undefined,
     }
 
     cy.stubAPIResponses()
@@ -218,7 +181,30 @@ describe(`${IC_COURT_RECORD_ROUTE}/:id`, () => {
 
     intercept(caseDataAddition)
 
-    cy.getByTestid('continueButton').should('not.exist')
+    cy.getByTestid('formFooter')
+      .children()
+      .getByTestid('infobox')
+      .should('exist')
+  })
+
+  it('should not allow users to continue if ruling is not set', () => {
+    const caseData = makeInvestigationCase()
+    const caseDataAddition: Case = {
+      ...caseData,
+      prosecutor: makeProsecutor(),
+      courtDate: '2021-12-16T10:50:04.033Z',
+      conclusion: faker.lorem.words(5),
+      decision: CaseDecision.ACCEPTING,
+      ruling: undefined,
+    }
+
+    cy.visit(`${IC_COURT_RECORD_ROUTE}/test_id_stadfest`)
+    intercept(caseDataAddition)
+
+    cy.getByTestid('formFooter')
+      .children()
+      .getByTestid('infobox')
+      .should('exist')
   })
 
   it('should navigate to the next step when all input data is valid and the continue button is clicked', () => {
@@ -230,6 +216,7 @@ describe(`${IC_COURT_RECORD_ROUTE}/:id`, () => {
       courtDate: '2021-12-16T10:50:04.033Z',
       decision: CaseDecision.ACCEPTING,
       conclusion: faker.lorem.words(5),
+      ruling: faker.lorem.words(5),
     }
 
     cy.stubAPIResponses()
