+++ apps/judicial-system/web/src/components/ConclusionDraft/ConclusionDraft.tsx
@@ -107,7 +107,6 @@ const ConclusionDraft: React.FC<Props> = (props) => {
       <RulingInput
         workingCase={workingCase}
         setWorkingCase={setWorkingCase}
-        isRequired
         rows={12}
       />
     </>
