--- apps/judicial-system/web/src/routes/Court/InvestigationCase/CourtRecord/CourtRecordForm.tsx
@@ -1,730 +1,0 @@
-import React, { useState } from 'react'
-import { useIntl } from 'react-intl'
-
-import {
-  Box,
-  GridColumn,
-  GridContainer,
-  GridRow,
-  Input,
-  RadioButton,
-  Text,
-  Tooltip,
-} from '@island.is/island-ui/core'
-import {
-  BlueBox,
-  CaseInfo,
-  CourtDocuments,
-  DateTime,
-  FormContentContainer,
-  FormFooter,
-  HideableText,
-  PdfButton,
-  TimeInputField,
-} from '@island.is/judicial-system-web/src/components'
-import {
-  Case,
-  CaseAppealDecision,
-  SessionArrangements,
-  User,
-} from '@island.is/judicial-system/types'
-import {
-  setAndSendDateToServer,
-  removeTabsValidateAndSet,
-  setAndSendToServer,
-  validateAndSendToServer,
-  validateAndSendTimeToServer,
-  validateAndSetTime,
-} from '@island.is/judicial-system-web/src/utils/formHelper'
-import { useCase } from '@island.is/judicial-system-web/src/utils/hooks'
-import {
-  formatDate,
-  formatRequestCaseType,
-} from '@island.is/judicial-system/formatters'
-import {
-  closedCourt,
-  icCourtRecord as m,
-  core,
-} from '@island.is/judicial-system-web/messages'
-import useDeb from '@island.is/judicial-system-web/src/utils/hooks/useDeb'
-import { isCourtRecordStepValidIC } from '@island.is/judicial-system-web/src/utils/validate'
-import * as Constants from '@island.is/judicial-system/consts'
-
-interface Props {
-  workingCase: Case
-  setWorkingCase: React.Dispatch<React.SetStateAction<Case>>
-  isLoading: boolean
-  user?: User
-}
-
-const CourtRecordForm: React.FC<Props> = (props) => {
-  const { workingCase, setWorkingCase, isLoading, user } = props
-  const [courtLocationEM, setCourtLocationEM] = useState<string>('')
-  const [
-    sessionBookingsErrorMessage,
-    setSessionBookingsMessage,
-  ] = useState<string>('')
-  const [courtDocumentEndEM, setCourtDocumentEndEM] = useState<string>('')
-
-  const { updateCase } = useCase()
-  const { formatMessage } = useIntl()
-
-  useDeb(workingCase, 'courtAttendees')
-  useDeb(workingCase, 'sessionBookings')
-  useDeb(workingCase, 'accusedAppealAnnouncement')
-  useDeb(workingCase, 'prosecutorAppealAnnouncement')
-  useDeb(workingCase, 'endOfSessionBookings')
-
-  return (
-    <>
-      <FormContentContainer>
-        <Box marginBottom={7}>
-          <Text as="h1" variant="h1">
-            Þingbók
-          </Text>
-        </Box>
-        <Box component="section" marginBottom={7}>
-          <CaseInfo workingCase={workingCase} userRole={user?.role} />
-        </Box>
-        <Box component="section" marginBottom={3}>
-          <BlueBox>
-            <Box marginBottom={3}>
-              <DateTime
-                name="courtStartDate"
-                datepickerLabel="Dagsetning þinghalds"
-                timeLabel="Þinghald hófst (kk:mm)"
-                maxDate={new Date()}
-                selectedDate={workingCase.courtStartDate}
-                onChange={(date: Date | undefined, valid: boolean) => {
-                  setAndSendDateToServer(
-                    'courtStartDate',
-                    date,
-                    valid,
-                    workingCase,
-                    setWorkingCase,
-                    updateCase,
-                  )
-                }}
-                blueBox={false}
-                required
-              />
-            </Box>
-            <Input
-              data-testid="courtLocation"
-              name="courtLocation"
-              tooltip={formatMessage(m.sections.courtLocation.tooltip)}
-              label={formatMessage(m.sections.courtLocation.label)}
-              value={workingCase.courtLocation || ''}
-              placeholder={formatMessage(m.sections.courtLocation.placeholder)}
-              onChange={(event) =>
-                removeTabsValidateAndSet(
-                  'courtLocation',
-                  event.target.value,
-                  ['empty'],
-                  workingCase,
-                  setWorkingCase,
-                  courtLocationEM,
-                  setCourtLocationEM,
-                )
-              }
-              onBlur={(event) =>
-                validateAndSendToServer(
-                  'courtLocation',
-                  event.target.value,
-                  ['empty'],
-                  workingCase,
-                  updateCase,
-                  setCourtLocationEM,
-                )
-              }
-              errorMessage={courtLocationEM}
-              hasError={courtLocationEM !== ''}
-              autoComplete="off"
-              required
-            />
-          </BlueBox>
-        </Box>
-        <Box component="section" marginBottom={8}>
-          <Box marginBottom={3}>
-            <HideableText
-              text={formatMessage(closedCourt.text)}
-              isHidden={workingCase.isClosedCourtHidden}
-              onToggleVisibility={(isVisible: boolean) =>
-                setAndSendToServer(
-                  'isClosedCourtHidden',
-                  isVisible,
-                  workingCase,
-                  setWorkingCase,
-                  updateCase,
-                )
-              }
-              tooltip={formatMessage(closedCourt.tooltip)}
-            />
-          </Box>
-          <Input
-            data-testid="courtAttendees"
-            name="courtAttendees"
-            label="Mættir eru"
-            value={workingCase.courtAttendees || ''}
-            placeholder="Skrifa hér..."
-            onChange={(event) =>
-              removeTabsValidateAndSet(
-                'courtAttendees',
-                event.target.value,
-                [],
-                workingCase,
-                setWorkingCase,
-              )
-            }
-            onBlur={(event) =>
-              updateCase(workingCase.id, { courtAttendees: event.target.value })
-            }
-            textarea
-            rows={7}
-            autoExpand={{ on: true, maxHeight: 300 }}
-          />
-        </Box>
-        <Box component="section" marginBottom={8}>
-          <Box marginBottom={2}>
-            <Text as="h3" variant="h3">
-              {formatMessage(m.sections.courtDocuments.header)}
-            </Text>
-          </Box>
-          <CourtDocuments
-            title={formatMessage(core.requestCaseType, {
-              caseType: formatRequestCaseType(workingCase.type),
-            })}
-            tagText={formatMessage(m.sections.courtDocuments.tag)}
-            tagVariant="darkerBlue"
-            text={formatMessage(m.sections.courtDocuments.text)}
-            caseId={workingCase.id}
-            selectedCourtDocuments={workingCase.courtDocuments ?? []}
-            onUpdateCase={updateCase}
-            setWorkingCase={setWorkingCase}
-            workingCase={workingCase}
-          />
-        </Box>
-        <Box component="section" marginBottom={8}>
-          <Box marginBottom={2}>
-            <Text as="h3" variant="h3">
-              {`${formatMessage(m.sections.sessionBookings.title)} `}
-              <Tooltip
-                text={formatMessage(m.sections.sessionBookings.tooltip)}
-              />
-            </Text>
-          </Box>
-          <Box marginBottom={3}>
-            <Input
-              data-testid="sessionBookings"
-              name="sessionBookings"
-              label={formatMessage(m.sections.sessionBookings.label)}
-              value={workingCase.sessionBookings || ''}
-              placeholder={formatMessage(
-                m.sections.sessionBookings.placeholder,
-              )}
-              onChange={(event) =>
-                removeTabsValidateAndSet(
-                  'sessionBookings',
-                  event.target.value,
-                  ['empty'],
-                  workingCase,
-                  setWorkingCase,
-                  sessionBookingsErrorMessage,
-                  setSessionBookingsMessage,
-                )
-              }
-              onBlur={(event) =>
-                validateAndSendToServer(
-                  'sessionBookings',
-                  event.target.value,
-                  ['empty'],
-                  workingCase,
-                  updateCase,
-                  setSessionBookingsMessage,
-                )
-              }
-              errorMessage={sessionBookingsErrorMessage}
-              hasError={sessionBookingsErrorMessage !== ''}
-              textarea
-              rows={16}
-              autoExpand={{ on: true, maxHeight: 600 }}
-              required
-            />
-          </Box>
-        </Box>
-        {workingCase.conclusion && (
-          <Box component="section" marginBottom={8}>
-            <Box marginBottom={2}>
-              <Text as="h3" variant="h3">
-                {formatMessage(m.sections.conclusion)}
-              </Text>
-            </Box>
-            <BlueBox>
-              <Text>{workingCase.conclusion}</Text>
-            </BlueBox>
-          </Box>
-        )}
-        <Box component="section" marginBottom={8}>
-          <Box marginBottom={2}>
-            <Text as="h3" variant="h3">
-              {formatMessage(m.sections.appealDecision.title)}
-            </Text>
-          </Box>
-          {workingCase.sessionArrangements ===
-            SessionArrangements.ALL_PRESENT && (
-            <Box marginBottom={3}>
-              <Text variant="h4" fontWeight="light">
-                {formatMessage(m.sections.appealDecision.disclaimer)}
-              </Text>
-            </Box>
-          )}
-          {workingCase.defendants && workingCase.defendants.length > 0 && (
-            <Box marginBottom={3}>
-              <BlueBox>
-                <Box marginBottom={2}>
-                  <Text as="h4" variant="h4">
-                    {formatMessage(m.sections.appealDecision.accusedTitle)}{' '}
-                    <Text as="span" color="red600" fontWeight="semiBold">
-                      *
-                    </Text>
-                  </Text>
-                </Box>
-                <Box marginBottom={2}>
-                  <GridRow>
-                    <GridColumn span="6/12">
-                      <RadioButton
-                        name="accused-appeal-decision"
-                        id="accused-appeal"
-                        label={formatMessage(
-                          workingCase.defendants.length > 1
-                            ? m.sections.appealDecision.multipleAccusedAppeal
-                            : m.sections.appealDecision.accusedAppeal,
-                        )}
-                        value={CaseAppealDecision.APPEAL}
-                        checked={
-                          workingCase.accusedAppealDecision ===
-                          CaseAppealDecision.APPEAL
-                        }
-                        onChange={() => {
-                          setWorkingCase({
-                            ...workingCase,
-                            accusedAppealDecision: CaseAppealDecision.APPEAL,
-                          })
-
-                          updateCase(workingCase.id, {
-                            accusedAppealDecision: CaseAppealDecision.APPEAL,
-                          })
-                        }}
-                        large
-                        backgroundColor="white"
-                      />
-                    </GridColumn>
-                    <GridColumn span="6/12">
-                      <RadioButton
-                        name="accused-appeal-decision"
-                        id="accused-accept"
-                        label={formatMessage(
-                          workingCase.defendants.length > 1
-                            ? m.sections.appealDecision.multipleAccusedAccept
-                            : m.sections.appealDecision.accusedAccept,
-                        )}
-                        value={CaseAppealDecision.ACCEPT}
-                        checked={
-                          workingCase.accusedAppealDecision ===
-                          CaseAppealDecision.ACCEPT
-                        }
-                        onChange={() => {
-                          setWorkingCase({
-                            ...workingCase,
-                            accusedAppealDecision: CaseAppealDecision.ACCEPT,
-                          })
-
-                          updateCase(workingCase.id, {
-                            accusedAppealDecision: CaseAppealDecision.ACCEPT,
-                          })
-                        }}
-                        large
-                        backgroundColor="white"
-                      />
-                    </GridColumn>
-                  </GridRow>
-                </Box>
-                <Box marginBottom={2}>
-                  <GridRow>
-                    <GridColumn span="7/12">
-                      <RadioButton
-                        name="accused-appeal-decision"
-                        id="accused-postpone"
-                        label={formatMessage(
-                          workingCase.defendants.length > 1
-                            ? m.sections.appealDecision.multipleAccusedPostpone
-                            : m.sections.appealDecision.accusedPostpone,
-                        )}
-                        value={CaseAppealDecision.POSTPONE}
-                        checked={
-                          workingCase.accusedAppealDecision ===
-                          CaseAppealDecision.POSTPONE
-                        }
-                        onChange={() => {
-                          setWorkingCase({
-                            ...workingCase,
-                            accusedAppealDecision: CaseAppealDecision.POSTPONE,
-                          })
-
-                          updateCase(workingCase.id, {
-                            accusedAppealDecision: CaseAppealDecision.POSTPONE,
-                          })
-                        }}
-                        large
-                        backgroundColor="white"
-                      />
-                    </GridColumn>
-                    <GridColumn span="5/12">
-                      <RadioButton
-                        name="accused-appeal-decision"
-                        id="accused-not-applicable"
-                        label={formatMessage(
-                          m.sections.appealDecision.accusedNotApplicable,
-                        )}
-                        value={CaseAppealDecision.NOT_APPLICABLE}
-                        checked={
-                          workingCase.accusedAppealDecision ===
-                          CaseAppealDecision.NOT_APPLICABLE
-                        }
-                        onChange={() => {
-                          setWorkingCase({
-                            ...workingCase,
-                            accusedAppealDecision:
-                              CaseAppealDecision.NOT_APPLICABLE,
-                          })
-
-                          updateCase(workingCase.id, {
-                            accusedAppealDecision:
-                              CaseAppealDecision.NOT_APPLICABLE,
-                          })
-                        }}
-                        large
-                        backgroundColor="white"
-                      />
-                    </GridColumn>
-                  </GridRow>
-                </Box>
-                <Input
-                  name="accusedAppealAnnouncement"
-                  data-testid="accusedAppealAnnouncement"
-                  label={formatMessage(
-                    m.sections.appealDecision.accusedAnnouncementLabel,
-                  )}
-                  value={workingCase.accusedAppealAnnouncement || ''}
-                  placeholder={formatMessage(
-                    workingCase.defendants.length > 1
-                      ? m.sections.appealDecision
-                          .multipleAccusedAnnouncementPlaceholder
-                      : m.sections.appealDecision
-                          .accusedAnnouncementPlaceholder,
-                  )}
-                  onChange={(event) =>
-                    removeTabsValidateAndSet(
-                      'accusedAppealAnnouncement',
-                      event.target.value,
-                      [],
-                      workingCase,
-                      setWorkingCase,
-                    )
-                  }
-                  onBlur={(event) =>
-                    validateAndSendToServer(
-                      'accusedAppealAnnouncement',
-                      event.target.value,
-                      [],
-                      workingCase,
-                      updateCase,
-                    )
-                  }
-                  textarea
-                  rows={7}
-                  autoExpand={{ on: true, maxHeight: 300 }}
-                />
-              </BlueBox>
-            </Box>
-          )}
-          <Box marginBottom={5}>
-            <BlueBox>
-              <Box marginBottom={2}>
-                <Text as="h4" variant="h4">
-                  {formatMessage(m.sections.appealDecision.prosecutorTitle)}{' '}
-                  <Text as="span" color="red400" fontWeight="semiBold">
-                    *
-                  </Text>
-                </Text>
-              </Box>
-              <Box marginBottom={2}>
-                <GridRow>
-                  <GridColumn span="6/12">
-                    <RadioButton
-                      name="prosecutor-appeal-decision"
-                      id="prosecutor-appeal"
-                      label={formatMessage(
-                        m.sections.appealDecision.prosecutorAppeal,
-                      )}
-                      value={CaseAppealDecision.APPEAL}
-                      checked={
-                        workingCase.prosecutorAppealDecision ===
-                        CaseAppealDecision.APPEAL
-                      }
-                      onChange={() => {
-                        setWorkingCase({
-                          ...workingCase,
-                          prosecutorAppealDecision: CaseAppealDecision.APPEAL,
-                        })
-
-                        updateCase(workingCase.id, {
-                          prosecutorAppealDecision: CaseAppealDecision.APPEAL,
-                        })
-                      }}
-                      large
-                      backgroundColor="white"
-                    />
-                  </GridColumn>
-                  <GridColumn span="6/12">
-                    <RadioButton
-                      name="prosecutor-appeal-decision"
-                      id="prosecutor-accept"
-                      label={formatMessage(
-                        m.sections.appealDecision.prosecutorAccept,
-                      )}
-                      value={CaseAppealDecision.ACCEPT}
-                      checked={
-                        workingCase.prosecutorAppealDecision ===
-                        CaseAppealDecision.ACCEPT
-                      }
-                      onChange={() => {
-                        setWorkingCase({
-                          ...workingCase,
-                          prosecutorAppealDecision: CaseAppealDecision.ACCEPT,
-                        })
-
-                        updateCase(workingCase.id, {
-                          prosecutorAppealDecision: CaseAppealDecision.ACCEPT,
-                        })
-                      }}
-                      large
-                      backgroundColor="white"
-                    />
-                  </GridColumn>
-                </GridRow>
-              </Box>
-              <Box marginBottom={2}>
-                <GridRow>
-                  <GridColumn span="7/12">
-                    <RadioButton
-                      name="prosecutor-appeal-decision"
-                      id="prosecutor-postpone"
-                      label={formatMessage(
-                        m.sections.appealDecision.prosecutorPostpone,
-                      )}
-                      value={CaseAppealDecision.POSTPONE}
-                      checked={
-                        workingCase.prosecutorAppealDecision ===
-                        CaseAppealDecision.POSTPONE
-                      }
-                      onChange={() => {
-                        setWorkingCase({
-                          ...workingCase,
-                          prosecutorAppealDecision: CaseAppealDecision.POSTPONE,
-                        })
-
-                        updateCase(workingCase.id, {
-                          prosecutorAppealDecision: CaseAppealDecision.POSTPONE,
-                        })
-                      }}
-                      large
-                      backgroundColor="white"
-                    />
-                  </GridColumn>
-                  <GridColumn span="5/12">
-                    <RadioButton
-                      name="prosecutor-appeal-decision"
-                      id="prosecutor-not-applicable"
-                      label={formatMessage(
-                        m.sections.appealDecision.prosecutorNotApplicable,
-                      )}
-                      value={CaseAppealDecision.NOT_APPLICABLE}
-                      checked={
-                        workingCase.prosecutorAppealDecision ===
-                        CaseAppealDecision.NOT_APPLICABLE
-                      }
-                      onChange={() => {
-                        setWorkingCase({
-                          ...workingCase,
-                          prosecutorAppealDecision:
-                            CaseAppealDecision.NOT_APPLICABLE,
-                        })
-
-                        updateCase(workingCase.id, {
-                          prosecutorAppealDecision:
-                            CaseAppealDecision.NOT_APPLICABLE,
-                        })
-                      }}
-                      large
-                      backgroundColor="white"
-                    />
-                  </GridColumn>
-                </GridRow>
-              </Box>
-              <Box>
-                <Input
-                  name="prosecutorAppealAnnouncement"
-                  data-testid="prosecutorAppealAnnouncement"
-                  label={formatMessage(
-                    m.sections.appealDecision.prosecutorAnnouncementLabel,
-                  )}
-                  value={workingCase.prosecutorAppealAnnouncement || ''}
-                  placeholder={formatMessage(
-                    m.sections.appealDecision.prosecutorAnnouncementPlaceholder,
-                  )}
-                  onChange={(event) =>
-                    removeTabsValidateAndSet(
-                      'prosecutorAppealAnnouncement',
-                      event.target.value,
-                      [],
-                      workingCase,
-                      setWorkingCase,
-                    )
-                  }
-                  onBlur={(event) =>
-                    validateAndSendToServer(
-                      'prosecutorAppealAnnouncement',
-                      event.target.value,
-                      [],
-                      workingCase,
-                      updateCase,
-                    )
-                  }
-                  textarea
-                  rows={7}
-                  autoExpand={{ on: true, maxHeight: 300 }}
-                />
-              </Box>
-            </BlueBox>
-          </Box>
-        </Box>
-        <Box component="section" marginBottom={5}>
-          <Box marginBottom={3}>
-            <Text as="h3" variant="h3">
-              {formatMessage(m.sections.endOfSessionBookings.title)}
-            </Text>
-          </Box>
-          <Box marginBottom={5}>
-            <Input
-              data-testid="endOfSessionBookings"
-              name="endOfSessionBookings"
-              label={formatMessage(m.sections.endOfSessionBookings.label)}
-              value={workingCase.endOfSessionBookings || ''}
-              placeholder={formatMessage(
-                m.sections.endOfSessionBookings.placeholder,
-              )}
-              onChange={(event) =>
-                removeTabsValidateAndSet(
-                  'endOfSessionBookings',
-                  event.target.value,
-                  [],
-                  workingCase,
-                  setWorkingCase,
-                )
-              }
-              onBlur={(event) =>
-                validateAndSendToServer(
-                  'endOfSessionBookings',
-                  event.target.value,
-                  [],
-                  workingCase,
-                  updateCase,
-                )
-              }
-              rows={16}
-              autoExpand={{ on: true, maxHeight: 600 }}
-              textarea
-            />
-          </Box>
-        </Box>
-        <Box marginBottom={5}>
-          <Box marginBottom={2}>
-            <Text as="h3" variant="h3">
-              Þinghald
-            </Text>
-          </Box>
-          <GridContainer>
-            <GridRow>
-              <GridColumn>
-                <TimeInputField
-                  onChange={(evt) =>
-                    validateAndSetTime(
-                      'courtEndTime',
-                      workingCase.courtStartDate,
-                      evt.target.value,
-                      ['empty', 'time-format'],
-                      workingCase,
-                      setWorkingCase,
-                      courtDocumentEndEM,
-                      setCourtDocumentEndEM,
-                    )
-                  }
-                  onBlur={(evt) =>
-                    validateAndSendTimeToServer(
-                      'courtEndTime',
-                      workingCase.courtStartDate,
-                      evt.target.value,
-                      ['empty', 'time-format'],
-                      workingCase,
-                      updateCase,
-                      setCourtDocumentEndEM,
-                    )
-                  }
-                >
-                  <Input
-                    data-testid="courtEndTime"
-                    name="courtEndTime"
-                    label="Þinghaldi lauk (kk:mm)"
-                    placeholder="Veldu tíma"
-                    autoComplete="off"
-                    defaultValue={formatDate(
-                      workingCase.courtEndTime,
-                      Constants.TIME_FORMAT,
-                    )}
-                    errorMessage={courtDocumentEndEM}
-                    hasError={courtDocumentEndEM !== ''}
-                    required
-                  />
-                </TimeInputField>
-              </GridColumn>
-            </GridRow>
-          </GridContainer>
-        </Box>
-        <Box marginBottom={10}>
-          <PdfButton
-            caseId={workingCase.id}
-            title={formatMessage(core.pdfButtonRulingShortVersion)}
-            pdfType="courtRecord"
-          />
-        </Box>
-      </FormContentContainer>
-      <FormContentContainer isFooter>
-        <FormFooter
-          previousUrl={`${Constants.IC_RULING_ROUTE}/${workingCase.id}`}
-          nextIsLoading={isLoading}
-          nextUrl={`${Constants.IC_CONFIRMATION_ROUTE}/${workingCase.id}`}
-          nextIsDisabled={!isCourtRecordStepValidIC(workingCase)}
-          hideNextButton={!workingCase.decision || !workingCase.conclusion}
-          infoBoxText={
-            !workingCase.decision || !workingCase.conclusion
-              ? formatMessage(m.nextButtonInfo)
-              : ''
-          }
-        />
-      </FormContentContainer>
-    </>
-  )
-}
-
-export default CourtRecordForm
